requirements.txt
mysql-connector-python==8.4.0
mysqlclient==2.2.4
pydantic==2.7.1
fastapi==0.111.0
python-dotenv~=1.0.1
uvicorn==0.29.0
starlette==0.37.2
python-jose
bcrypt